better scripts...
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@201 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ deploy.sh
@@ -9,8 +9,16 @@ fi
 
 cp dist/*.war $JBOSS_HOME/deploy
 cp dist/*.jar $JBOSS_HOME/deploy
+echo Deployed jar- and war-files in $JBOSS_HOME/deploy
 
-cp lib/jce-jdk13-111.jar $JBOSS_HOME/lib/ext
-cp lib/ldap.jar $JBOSS_HOME/lib/ext
+if ! [ -f $JBOSS_HOME/lib/ext/jce-jdk13-111.jar ]
+then
+  cp lib/jce-jdk13-111.jar $JBOSS_HOME/lib/ext
+  echo Copied jce-jdk13-111.jar to $JBOSS_HOME/lib/ext. JBoss must be restared.
+fi
+if ! [ -f $JBOSS_HOME/lib/ext/ldap.jar ]
+then
+  cp lib/ldap.jar $JBOSS_HOME/lib/ext
+  echo Copied ldap.jar to $JBOSS_HOME/lib/ext. JBoss must be restared.
+fi
 
-#echo jce-jdk13-111.jar and ldap.jar  must be placed in jboss/lib/ext.
